# perf(polys): speed up AlgebraicField.from_sympy