@oscarbenjamin 2024-08-01
> With master this would take a lot longer...

I estimate that it would take about an hour. So the change here makes it about 1000x faster for the case shown.