@oscarbenjamin 2024-08-20
It seems that this fails for the PR:
```python
In [3]: K = QQ.algebraic_field(exp(I*pi/7))

In [4]: a = K.from_sympy(cos(2*pi/7)-I*sin(2*pi/7))
---------------------------------------------------------------------------
IsomorphismFailed                         Traceback (most recent call last)
File ~/current/active/sympy/sympy/polys/domains/algebraicfield.py:583, in _make_input_converter.<locals>.isomorphism_converter(a)
    582 try:
--> 583     res = K(to_number_field(a, K.ext).native_coeffs())
    584 except (NotAlgebraic, IsomorphismFailed) as exc:

File ~/current/active/sympy/sympy/polys/numberfields/subfield.py:506, in to_number_field(extension, theta, gen, alias)
    505 else:
--> 506     raise IsomorphismFailed(
    507         "%s is not in a subfield of %s" % (root, theta.root))

IsomorphismFailed: I is not in a subfield of exp(I*pi/7)

The above exception was the direct cause of the following exception:

CoercionFailed                            Traceback (most recent call last)
Cell In[4], line 1
----> 1 a = K.from_sympy(cos(2*pi/7)-I*sin(2*pi/7))

File ~/current/active/sympy/sympy/polys/domains/algebraicfield.py:365, in AlgebraicField.from_sympy(self, a)
    363 if conv is None:
    364     conv = self._input_converter = _make_input_converter(self)
--> 365 return conv(a)

File ~/current/active/sympy/sympy/polys/domains/algebraicfield.py:598, in _make_input_converter.<locals>.converter(a)
    596     result = K.zero
    597     for ai in a.args:
--> 598         result += converter(ai)
    599     return result
    600 elif a.is_Mul:

File ~/current/active/sympy/sympy/polys/domains/algebraicfield.py:603, in _make_input_converter.<locals>.converter(a)
    601     result = K.one
    602     for ai in a.args:
--> 603         result *= converter(ai)
    604     return result
    605 elif a.is_Pow and a.exp.is_Integer:

File ~/current/active/sympy/sympy/polys/domains/algebraicfield.py:612, in _make_input_converter.<locals>.converter(a)
    610 res = cache.get(a)
    611 if res is None:
--> 612     res = cache[a] = isomorphism_converter(a)
    613 return res

File ~/current/active/sympy/sympy/polys/domains/algebraicfield.py:585, in _make_input_converter.<locals>.isomorphism_converter(a)
    583     res = K(to_number_field(a, K.ext).native_coeffs())
    584 except (NotAlgebraic, IsomorphismFailed) as exc:
--> 585     raise CoercionFailed(
    586         "%s is not a valid algebraic number in %s" % (a, K)) from exc
    587 return res

CoercionFailed: I is not a valid algebraic number in QQ<exp(I*pi/7)>
```
On master it gives:
```python
In [9]: K = QQ.algebraic_field(exp(I*pi/7))

In [10]: a = K.from_sympy(cos(2*pi/7)-I*sin(2*pi/7))

In [11]: N(-exp(I*pi/7)**5)
Out[11]: 0.623489801858733 - 0.78183148246803⋅ⅈ

In [12]: N(cos(2*pi/7)-I*sin(2*pi/7))
Out[12]: 0.623489801858733 - 0.78183148246803⋅ⅈ

In [13]: K.from_sympy(I)
...
IsomorphismFailed: I is not in a subfield of exp(I*pi/7)
```
So the bug is trying to convert each part of the expression separately into the number field.

Is it correct that `I` is not in this field?
```python
In [16]: factor(x**2 + 1, extension=exp(I*pi/7))
Out[16]: 
 2    
x  + 1
```
Maybe we can only split out rational factors from a product...

More generally though there is no guarantee that real and imaginary parts of an algebraic number belong to the same field although it works in this case:
```python
In [17]: K.from_sympy(cos(2*pi/7))
Out[17]: ANP([-1/2, 0, 0, 1/2, 0, 0], [1, -1, 1, -1, 1, -1, 1], QQ)

In [18]: K.from_sympy(sin(2*pi/7))
---------------------------------------------------------------------------
IsomorphismFailed 
```
I suppose that when decomposing an Add of Muls we need to make a strictly Q-linear decomposition...