@oscarbenjamin 2024-08-20
> I suppose that when decomposing an Add of Muls we need to make a strictly Q-linear decomposition...

Actually there is no guarantee that we can split an Add or a Mul into terms or factors here...