@oscarbenjamin 2024-08-20
Marking this as draft. More thought is needed.

Maybe using `ex=True` and handling the primitive element expressions is the right thing and other expressions just have to go through the whole `to_number_field` dance.